load assets on any url and added animations when displaying Deviations
Only assets available on root-path will work with this technique.

in a request GET /whatever/asset.png
We check in the route if the last part of the url, 'asset.png',
is present in the asset folder.
If it is, send the file.
If not, goto next() handler.

A limitation with this approach is only assets that are available on / (root) can be
detected without a lot of extra work.